% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McSweeney_Porter.r
\name{McSweeney_Porter}
\alias{McSweeney_Porter}
\title{McSweeney and Porter Method for Nonparametric ANCOVA}
\usage{
McSweeney_Porter(formula, data)
}
\arguments{
\item{formula}{An object of class "formula": a symbolic description of the 
model to be fitted. The structure should be `response ~ covariate1 + ... + group`.
The grouping variable should be the last term in the formula.}

\item{data}{A data frame containing the variables specified in the formula.
All variables used in the model should be complete; missing values are not allowed.}
}
\value{
A list containing the following components:
\describe{
  \item{regression_equation_covariate}{The summary of the model with only covariates.}
  \item{regression_equation_covariate_group}{The summary of the model with covariates and group main effects.}
  \item{regression_equation_interaction}{The summary of the model including the interaction term.}
  \item{group_effect}{The result of an ANOVA test for group effect.}
  \item{interaction_effect}{The result of an ANOVA test for interaction effect between group and covariate variables.}
  \item{data}{The original data frame with added rank columns.}
}
}
\description{
Performs rank-based ANCOVA with and without an interaction term between the covariates and the group.
}
\examples{
# 1. Create a sample data frame
data <- data.frame(
  group = c(1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3),
  response = c(16, 60, 82, 126, 137, 44, 67, 87, 100, 142, 17, 28, 105, 149, 160),
  covariate1 = c(26, 10, 42, 49, 55, 21, 28, 5, 12, 58, 1, 19, 41, 48, 35),
  covariate2 = c(12, 21, 24, 29, 34, 17, 2, 40, 38, 36, 8, 1, 9, 28, 16)
)

# 2. Run the McSweeney and Porter method
results <- McSweeney_Porter(
  formula = response ~ covariate1 + covariate2 + group,
  data = data
)

# 3. View the results
print(results)
print(results$group_effect)
print(results$interaction_effect)

}
\references{
McSweeney, M., & Porter, A. (1971). Small sample properties of nonparametric index of response and rank analysis of covariance. \emph{Omega}, 16.

Olejnik, S. F., & Algina, J. (1985). A review of nonparametric alternatives to analysis of covariance. \emph{Journal of Educational Research}, 9(1), 51-83.
}
