% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_npANCOVA.r
\name{pairwise_npANCOVA}
\alias{pairwise_npANCOVA}
\title{Pairwise Post Hoc Comparisons for Nonparametric ANCOVA}
\usage{
pairwise_npANCOVA(
  formula,
  data,
  method = c("Burnett_Barr", "Harwell_Serlin", "Puri_Sen_UVCM", "Puri_Sen_BVCM", "Quade",
    "Shirley", "McSweeney_Porter", "Hettmansperger_McKean"),
  p_adjust_methods = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr"),
  alpha = 0.05
)
}
\arguments{
\item{formula}{An object of class "formula": a symbolic description of the model to be fitted.
The structure should be ‘response ~ covariate1 + ... + group‘. 
The grouping variable should be the last term in the formula.}

\item{data}{A data frame containing the variables specified  in the model. All variables 
used in the model should be complete; missing values are not allowed.}

\item{method}{A character string specifying the nonparametric ANCOVA method.
Should be one of \code{"Burnett_Barr"}, \code{"Harwell_Serlin"},
\code{"Puri_Sen_UVCM"}, \code{"Puri_Sen_BVCM"}, \code{"Quade"}, \code{"Shirley"},
\code{"McSweeney_Porter"}, or \code{"Hettmansperger_McKean"}.}

\item{p_adjust_methods}{A character vector specifying one or more p-value
adjustment methods to be applied to the raw pairwise p-values using
\code{\link[stats]{p.adjust}}.}

\item{alpha}{A numeric significance level used to determine whether adjusted
p-values are statistically significant.}
}
\value{
A list containing:
\describe{
  \item{method}{The selected nonparametric ANCOVA method.}
  \item{formula}{The model formula used in the analysis.}
  \item{alpha}{The significance level used for adjusted comparisons.}
  \item{comparisons}{A data frame containing pairwise group comparisons,
  raw (unadjusted) p-values, adjusted p-values, and significance decisions.}
}
}
\description{
Performs pairwise post hoc comparisons for nonparametric ANCOVA
methods by fitting the selected nonparametric method separately to each
pair of groups. For every pairwise comparison, the function extracts the raw
p-value from the fitted model and computes multiple-testing adjusted p-values.
}
\details{
All possible pairwise combinations of group levels are generated, and nonparametric
ANCOVA method is fitted to each two-group subset using the selected method.

Raw p-values are extracted from each method and adjusted using one or more
methods supported by \code{\link[stats]{p.adjust}}. The adjusted p-values are
returned together with significance decisions based on the specified significance
level \code{alpha}.

Supported methods are:
\code{"Burnett_Barr"},
\code{"Harwell_Serlin"},
\code{"Puri_Sen_UVCM"},
\code{"Puri_Sen_BVCM"},
\code{"Quade"},
\code{"Shirley"},
\code{"McSweeney_Porter"},
and \code{"Hettmansperger_McKean"}.
}
\examples{
data <- data.frame(
  group = c(1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3),
  response = c(16, 60, 82, 126, 137, 44, 67, 87, 100, 142, 17, 28, 105, 149, 160),
  covariate1 = c(26, 10, 42, 49, 55, 21, 28, 5, 12, 58, 1, 19, 41, 48, 35)
)

# Burnett_Barr
pairwise_npANCOVA(
  response ~ covariate1 + group,
  data = data,
  method = "Burnett_Barr"
)

# Harwell_Serlin
pairwise_npANCOVA(
  response ~ covariate1 + group,
  data = data,
  method = "Harwell_Serlin"
)

# Puri_Sen_UVCM
pairwise_npANCOVA(
  response ~ covariate1 + group,
  data = data,
  method = "Puri_Sen_UVCM"
)

# Puri_Sen_BVCM
pairwise_npANCOVA(
  response ~ covariate1 + group,
  data = data,
  method = "Puri_Sen_BVCM"
)


# Quade
pairwise_npANCOVA(
  response ~ covariate1 + group,
  data = data,
  method = "Quade"
)

# Shirley
pairwise_npANCOVA(
  response ~ covariate1 + group,
  data = data,
  method = "Shirley"
)

# McSweeney_Porter
pairwise_npANCOVA(
  response ~ covariate1 + group,
  data = data,
  method = "McSweeney_Porter"
)

# Hettmansperger_McKean
pairwise_npANCOVA(
  response ~ covariate1 + group,
  data = data,
  method = "Hettmansperger_McKean"
)

}
