% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_npANCOVA.r
\name{summary_npANCOVA}
\alias{summary_npANCOVA}
\title{Summarize Nonparametric ANCOVA Results}
\usage{
summary_npANCOVA(object)
}
\arguments{
\item{object}{A fitted result object returned by one
of the supported methods: \code{Burnett_Barr()},
\code{Harwell_Serlin()}, \code{Hettmansperger_McKean()},
\code{McSweeney_Porter()}, \code{Puri_Sen_BVCM()},
\code{Puri_Sen_UVCM()}, \code{Quade()}, or \code{Shirley()}.}
}
\value{
Prints a formatted summary of the fitted object containing the following components:
\describe{
  \item{statistics}{The test statistic.}
  \item{df}{The degrees of freedom for the test.}
  \item{p_value}{The p-value of the test.}
}
}
\description{
This function formats and prints inferential results obtained from a fitted
nonparametric ANCOVA object.
}
\details{
The function uses the \code{method_name} stored in \code{object} and prints the
appropriate inferential summary according to the detected nonparametric
ANCOVA method.

\itemize{
  \item For \code{"Burnett_Barr"} and \code{"Quade"},
  the function prints the F-statistic, numerator and denominator degrees of freedom, and p-value.

  \item For \code{"Hettmansperger_McKean"}, \code{"Harwell_Serlin"}, \code{"Puri_Sen_BVCM"}, and
  \code{"Puri_Sen_UVCM"}, the function prints the chi-squared test statistic, degrees
  of freedom, and p-value.

  \item For \code{"McSweeney_Porter"} and \code{"Shirley"}, both group and
  interaction effects are printed when available.
}
}
\examples{
data <- data.frame(
  group = c(1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3),
  response = c(16, 60, 82, 126, 137, 44, 67, 87, 100, 142, 17, 28, 105, 149, 160),
  covariate1 = c(26, 10, 42, 49, 55, 21, 28, 5, 12, 58, 1, 19, 41, 48, 35),
  covariate2 = c(12, 21, 24, 29, 34, 17, 2, 40, 38, 36, 8, 1, 9, 28, 16)
)

formula_one_cov <- response ~ covariate1 + group
formula_two_cov <- response ~ covariate1 + covariate2 + group

# Method requiring exactly one covariate
fit_bb <- Burnett_Barr(formula_one_cov, data = data)
summary_npANCOVA(fit_bb)

# Methods allowing multiple covariates
fit_hs <- Harwell_Serlin(formula_two_cov, data = data)
summary_npANCOVA(fit_hs)

fit_hm <- Hettmansperger_McKean(formula_two_cov, data = data)
summary_npANCOVA(fit_hm)

fit_mp <- McSweeney_Porter(formula_two_cov, data = data)
summary_npANCOVA(fit_mp)

fit_psmb <- Puri_Sen_BVCM(formula_two_cov, data = data)
summary_npANCOVA(fit_psmb)

fit_psmu <- Puri_Sen_UVCM(formula_two_cov, data = data)
summary_npANCOVA(fit_psmu)

fit_quade <- Quade(formula_two_cov, data = data)
summary_npANCOVA(fit_quade)

fit_shirley <- Shirley(formula_two_cov, data = data)
summary_npANCOVA(fit_shirley)

}
\seealso{
\code{\link{Burnett_Barr}},
\code{\link{Quade}},
\code{\link{Hettmansperger_McKean}},
\code{\link{Harwell_Serlin}},
\code{\link{McSweeney_Porter}},
\code{\link{Shirley}},
\code{\link{Puri_Sen_BVCM}},
\code{\link{Puri_Sen_UVCM}},
}
